% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdiabaticTandLWC.R
\name{AdiabaticTandLWC}
\alias{AdiabaticLWC}
\alias{AdiabaticTandLWC}
\title{AdiabaticTandLWC}
\usage{
AdiabaticTandLWC(.Pbase, .Tbase, .Pobs, .Ebase = 0, .lwc = 0)
}
\arguments{
\item{.Pbase}{Starting-point pressure, hPa}

\item{.Tbase}{Starting-point temperature, degC}

\item{.Pobs}{Pressure at the observation level, hPa (can be length>1)}

\item{.Ebase}{Optional vapor pressure at starting point; the default
is zero and in that case the calculation uses the equilibrium vapor 
pressure at .Tbase}

\item{.lwc}{Liquid water content (g/m^3) at starting point (default 0)}
}
\value{
A data.frame containing two columns, Tobs and ALWC, which are the
(possibly length>1) values of temperature [degC] and liquid water content
[g/m^3] corresponding to the pressures .pObs .
}
\description{
Calculate the temperature and liquid water content for adiabatic ascent.
}
\details{
For reversible ascent, this function calculates the final 
temperature and liquid water content given an initial pressure, 
temperature, and liquid water content. The starting point will often 
be the cloud base, and .lwc will be zero, but the routine doesn't 
require this and if given a subsaturated vapor pressure will
calculate reversible ascent below cloud base and produce zero LWC
for final pressures below cloud base. If the optional vapor pressure
at the starting point is zero or omittedthe equilibrium vapor pressure at the
starting-point temperature will be used.
}
\examples{
\dontrun{R <- AdiabaticTandLWC (900., 20., c(900,800,700,600,500,400)) ; print(R)}
}
\author{
William Cooper
}

