% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AirTemperature.R
\name{AirTemperature}
\alias{AirTemperature}
\title{AirTemperature}
\usage{
AirTemperature(RT, P, Q, E = 0)
}
\arguments{
\item{RT}{A numeric representing the recovery temperature in deg. C}

\item{P}{A numeric representing the ambient pressure in hPa}

\item{Q}{A numeric representing the dynamic pressure in hPa}

\item{E}{An optional numeric representing the water vapor pressure in hPa}
}
\value{
The ambient air temperature in deg.C
}
\description{
Calculates the ambient air temperature
}
\details{
Finds the air temperature from measurements of recovery temperature and ambient and dynamic pressure, with optional correction for humidity
}
\examples{
AT <- AirTemperature (10., 700., 50.)
}
\author{
William Cooper
}

