% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binStats.R
\name{binStats}
\alias{binStats}
\title{binStats}
\usage{
binStats(.d, bins = 20, xlow = min(.d[, 2], na.rm = TRUE),
  xhigh = max(.d[, 2], na.rm = TRUE))
}
\arguments{
\item{.d}{A data-frame containing at least two columns, the first holding
values to partition and summarize and the second the variable used for
defining bins. Values from the first will be grouped according to the
bins in the second and then a mean and standard deviation will be
calculated for each group.}

\item{bins}{(default 20)}

\item{xlow}{The lower limit for bins. The default is the minimum of the
second variable.}

\item{xhigh}{The upper limit for bins. The default is the maximum of the
second variable.}
}
\value{
A dataframe of dimension c(bins,4) that contains, for each bin,
'xc' = the center coordinate of the bin, 'ybar' = the mean value for the bin
(or NA if there are no members of the group), 'sigma' = the standard
deviation for the bin, or NA for no members of the group, and 'N' =
the number of values in the bin.
}
\description{
bin and calculate mean and st.dev. for each bin
}
\details{
Given two corresponding variables in a data frame, defines
bins in the second, bins corresponding values of the first, and
calculates the mean and standard deviation of the first in each bin.
The purpose is to provide values for error bars, e.g., for geom_errorbar
where ymin=ybar-sigma and ymax=ybar+sigma.
}
\examples{
\dontrun{E <- binStats(Data[, c("V1", "V2"), bins=15)}
}
\author{
William Cooper
}

