% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotentialTemperatures.R
\name{BoltonEquivalentPotentialTemperature}
\alias{BoltonEquivalentPotentialTemperature}
\title{BoltonEquivalentPotentialTemperature}
\usage{
BoltonEquivalentPotentialTemperature(P, AT, E = 0)
}
\arguments{
\item{P}{A numeric representing ambient pressure in hPa}

\item{AT}{A numeric representing air temperature in deg. C}

\item{E}{A numeric representing water vapor pressure in hPa. Defaults to zero,
and for that case the calculation uses the equilibrium vapor pressure at AT.}
}
\value{
A numeric representing the Bolton-equation result for the pseudo-adiabatic equivalent potential temperature in kelvin.
}
\description{
The Bolton form of the pseudo-adiabatic equivalent potential temperature
}
\details{
The Bolton form of the pseudo-adiabatic equivalent potential temperature is calculated according to the formula of Davies-Jones (2009); see ProcessingAlgorithms.pdf
}
\examples{
THETAB <- BoltonEquivalentPotentialTemperature (700., 10., 9.) 
}
\author{
William Cooper
}
