% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HeadingCorrection.R
\name{CorrectHeading}
\alias{CorrectHeading}
\title{CorrectHeading}
\usage{
CorrectHeading(.data = Data, .span = 21, .default = -0.08)
}
\arguments{
\item{.data}{A data.frame containing measurements of body accelerations (called
BLATA, BLONGA, BNORMA), ground-speed components from GPS (GGVEW, GGVNS), attitude
angles in units of degrees (PITCH, ROLL, THDG, the latter the heading relative to
true north), and the latitude and altitude (LAT and GGALT), the latter two for
finding the local acceleration of gravity. If variables named PITCHC, ROLLC, and
LATC are present those will be used instead. The first two might be produced by
prior use of the CorrectPitch() function.}

\item{.span}{The number of points (default 21) used for estimating the accelerations
by differentiation and for smoothing the results.}

\item{.default}{The default value to use if the routine does not find enough
qualified turns to develop a valid estimate of the heading error. The default
is -0.08 deg., as applies to the GV data for DEEPWAVE.}
}
\value{
A vector of the same length as the supplied data.frame that gives an
estimate of the error in heading, the negative of the correction needed.
}
\description{
Calculate an estimate of the sequence of errors in heading.
}
\details{
For a data.frame supplied with variables representing the attitude angles,
body accelerations, GPS-measured ground velocities, and altitude and latitude,
estimate the error in the heading via comparison of the measured body accelerations,
transformed to the Earth-relative frame, to the accelerations determined via
differentiation of the GPS-measured ground-speed components. Find the correction
to heading that provides the best match between these two sources of acceleration.
The result is the estimated error, so this should be **subtracted** from the
measured value to get the corrected value. Results are sensitive to timing
delays among the variables, so these should be corrected prior to calling this routine.
}
\examples{
\dontrun{HeadingCorrection <- CorrectHeading (DataFrame)}
}
\author{
William Cooper
}

