% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PitchCorrection.R
\name{CorrectPitch}
\alias{CorrectPitch}
\title{CorrectPitch}
\usage{
CorrectPitch(D)
}
\arguments{
\item{D}{a data.frame containing at least these variables:
VNS, VEW, GGVNS, GGVEW, LAT, GGALT, THDG, PITCH, ROLL}
}
\value{
PitchError -- the estimated error in the pitch [deg], which
should be subtracted from PITCH to get the corrected value.
}
\description{
Calculate a correction to pitch based on the Schuler oscillation.
}
\details{
Uses measurements of the ground-speed error (as determined by comparison
to GPS-measured ground-speed components) to estimate the error in pitch. This
works best if long segments of flight are included in the input data.frame so
there will be sufficient time to determine the Schuler oscillation well. For
high-rate files, calculations are based on 1-Hz data and interpolated because
the pitch correction is smoothed to be slowly varying over periods of several
minutes.
}
\examples{
\dontrun{PITCHC <- PITCH - CorrectPitch(D)}
}
\author{
William Cooper
}

