% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctT.R
\name{correctT}
\alias{correctT}
\title{correctT}
\usage{
correctT(.data)
}
\arguments{
\item{.data}{A data.frame containing measurements of air temperature and
also airspeed (TASX), Mach Number (either MACHX or XMACH2), ambient pressure
(PSXC) and water vapor pressure (either EWX or EDPC),}
}
\value{
The supplied data.frame with new variables added that represent 
correction for time response (added with suffix "C") and replacement of the
conventional dynamic-heating correction by a version filtered to match
the time response of the sensor (added with suffix "F").
}
\description{
Calculate time-response-corrected and dynamic-heating-filtered
temperature variables for each temperature measurement in a supplied data.frame.
}
\details{
For a data.frame supplied with some variables representing the air
temperature, corrected values for each are calculated and added to the supplied
data.frame. This routine is specific to the temperature sensors on the
NSF/NCAR GV and C-130, either unheated Rosemount sensors or heated HARCO
sensors, and it depends on conventional attributes in the NCAR/EOL/RAF
netCDF files to determine how to handle each sensor. The supplied data.frame 
must contain additional variables as specified for the first parameter below.
}
\author{
William Cooper
}
