% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDirectory.R
\name{DataDirectory}
\alias{DataDirectory}
\title{DataDirectory}
\usage{
DataDirectory()
}
\value{
Directory The path to the local data directory, or NA if none found
}
\description{
Returns the location of the local data directory.
}
\details{
This is intended to make it easier to write programs that work on different computers where the netCDF files by convention are stored in different locations. It returns the local location, or NA if none of the standard locations are found.
}
\examples{
Directory <- DataDirectory ()
}
\author{
William Cooper
}

