% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFileInfo.R
\name{DataFileInfo}
\alias{DataFileInfo}
\alias{dataFileInfo}
\title{DataFileInfo}
\usage{
DataFileInfo(fileLocation, LLrange = TRUE)
}
\arguments{
\item{fileLocation}{A complete address to the netCDF data file.}

\item{LLrange}{Logical, if TRUE, range in latitude and longitude for the
flight is included in the returned list. Set FALSE to get a faster return
in case the range in positions is not needed.}
}
\value{
A list containing named characteristics of the file. For example, 
in the example below the variables in the file are returned in FI$Variables.
See names(FI) for a list of characteristics that are returned.
}
\description{
Brief information on the contents of a netCDF data file
}
\details{
Reads the flight number, project, date/time, position, and variables from
the netCDF file and returns a list of properties. Alternately, produces similar
information for a saved (.Rdata-format) data.frame produced by reading such a
netCDF file.
}
\examples{
FI <-DataFileInfo (sprintf ("\%s/extdata/RAFdata.nc", path.package ("Ranadu")))
}
\author{
William Cooper
}
