% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DemingFit}
\alias{DemingFit}
\title{Deming Fit}
\usage{
DemingFit(.x, .y, .sdx = 1, .sdy = 1)
}
\arguments{
\item{.x}{A numeric vector of measurements from one source.}

\item{.y}{A numeric vector of measurements from another source.}

\item{.sdx}{A numeric vector specifying the standard
measurement uncertainty for the first set of measurements. Default: 1}

\item{.sdy}{Like .sdx but for the second set of measurements.}
}
\value{
c(a, b, rms) for a fit of the form y = a + b * x, or NULL if no fit is possible (e.g., too few points or zero correlation)
}
\description{
Fit a line to data that minimizes the least-squared
distance of the data points from the line.
}
\details{
See the detailed description in DemingFit.pdf
}
\examples{
DemingFit ((1:5+rnorm(5,0,0.5)), (1:5+rnorm(5,0,0.5)))
}
\author{
William Cooper
}

