% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{detrend}
\alias{detrend}
\alias{deTrend}
\title{detrend}
\usage{
detrend(.data, silent = TRUE)
}
\arguments{
\item{.data}{A data.frame containing variable 'Time' as the first column and
a second-column from which to remove the mean and trend.}

\item{silent}{Prints mean and trend only if set FALSE. Default is TRUE.}
}
\value{
A modified version of data[, 2] only, with mean and trend removed.
}
\description{
Simple removal of mean and trend from a time series.
}
\details{
Input must be a data.frame where the first column is 'Time'
and the second is the variable from which to remove the mean and trend.
Removing the mean often provides less numerical inaccuracy, and removing
a trend (for spectral analysis) avoids the effect of a ramp function
on an fft-produced spectrum for which the time series is assumed repetitive.
For this calculation, NA values are replaced by interpolation or, where that
is not possible, by zeroes.
}
\examples{
AT <- detrend (RAFdata[, c('Time', 'ATX')])
}
\author{
William Cooper
}
