% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PotentialTemperatures.R
\name{EquivalentPotentialTemperature}
\alias{EquivalentPotentialTemperature}
\title{EquivalentPotentialTemperature}
\usage{
EquivalentPotentialTemperature(P, AT, E = 0)
}
\arguments{
\item{P}{A numeric representing ambient pressure in hPa}

\item{AT}{A numeric representing air temperature in deg. C}

\item{E}{A numeric representing water vapor pressure in hPa. Defaults to zero,
and for that case the calculation uses the equilibrium vapor pressure at AT.}
}
\value{
A numeric representing the pseudo-adiabatic equivalent potential temperature in kelvin.
}
\description{
The pseudo-adiabatic equivalent potential temperature
}
\details{
The pseudo-adiabatic equivalent potential temperature is calculated according to the formula of Davies-Jones (2009); see ProcessingAlgorithms.pdf
}
\examples{
THETAP <- EquivalentPotentialTemperature (700., 10., 9.)
}
\author{
William Cooper
}

