% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallv.R
\name{fallv}
\alias{fallv}
\title{fallv}
\usage{
fallv(d, rho = 1000, T = 0, p = 1013.25)
}
\arguments{
\item{d}{The droplet diameter [m]; default=20e-6}

\item{rho}{Density of the hydrometeor. Default=1.e3}

\item{T}{temperature [deg.C]}

\item{p}{pressure [hPa]}
}
\value{
The terminal fall speed [m/s]
}
\description{
Provides the fall speeds of spherical hydrometeors
}
\examples{
fallv (100e-6)
}
\author{
William Cooper
}
