% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoPotHeight.R
\name{GeoPotHeight}
\alias{GeoPotHeight}
\title{GeoPotHeight}
\usage{
GeoPotHeight(.latitude, .H = 0, .geoid = 0)
}
\arguments{
\item{.latitude}{The latitude (or vector of latitudes) in units of degrees N.}

\item{.H}{The altitude in meters above the geoid representing mean sea level, or a vector
containing such quantities. The GPS-provided variable "GGALT" is appropriate.}

\item{.geoid}{The height of the geoid above the WGS84 reference ellipse. This usually
introduces a small correction to the results obtained with .geoid equal to zero. The
variable GGEOIDHT in RAF-supplied data files is appropriate to use here.}
}
\value{
A variable or vector of variables representing geopotential height [m].
}
\description{
Provides the geopotential height.
}
\details{
The formula used incorporates the dependence of gravity on latitude and on 
altitude above the WGS-84 geoid, as represented by the Somigliana factor and a
Taylor expansion of the dependence on height. See the code for "Gravity.R" for
more details regarding the coefficients used.
}
\author{
William Cooper
}
