% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getarg.R
\name{getarg}
\alias{getarg}
\alias{getarg,}
\alias{getargs}
\title{getarg}
\usage{
getarg(V)
}
\arguments{
\item{V}{A variable list specified in any of the ways specified above.}
}
\description{
A convenience function providing flexibility in specification of function arguments.
}
\details{
The function returns a character-vector of argument names for an input name that can
have any of these forms: getarg(c(ATX, DPXC)), getarg(names(Data)), getarg(names(Data)[2:3]), 
getarg(nm[2:4]), getarg(ATX), getarg('ATX'), getarg(c('ATX', 'PSXC')), getarg(c(ATX)). 
If called with argument Data$ATX, returns c('Data', 'ATX'); for Data[ir, ATX], returns
c('Data', 'ir', 'ATX'). These forms are used in some Ranadu functions, notably VSpec(),
for more flexibility in the arguments that can be supplied.
}
\examples{
getarg(c(RTH1, RTH2))

}
\author{
William Cooper
}
