% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TellAbout.R
\name{getAttributes}
\alias{getAttributes}
\alias{getAttributes,GetAttributes}
\title{getAttributes}
\usage{
getAttributes(dname = Data, vname = NULL)
}
\arguments{
\item{dname}{The name of the data.frame. Alternately, a data.frame column.}

\item{vname}{The name of a variable in the netCDF file. If provided, this should
be specified as a string variable (e.g., "TASX"). If omitted, the attributes of
the first parameter will be returned.}
}
\value{
A list of the attributes and associated values
}
\description{
List the netCDF attributes for a specified variable or data.frame
}
\details{
Prints the netCDF attributes and their values, and returns
a text vector with the names of the attributes, for a specified
data.frame. For variable attributes, the name of the variable should be
specified as a string. For global attributes, omit the 'vname' parameter.
This function is designed to work with data.frames generated by Ranadu::getNetCDF.R,
and for such data.frames it will list attributes as in the original netCDF file from
which the data.frame was generated.
}
\examples{
\dontrun{attr <- GetAttributes (Data=getNetCDF(fname, "TASX"), "TASX")}
\dontrun{attr <- GetAttributes (Data$TASX)}
}
\author{
William Cooper
}

