% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRoutines.R
\name{getIndex}
\alias{getIndex}
\title{getIndex}
\usage{
getIndex(Time, HHMMSS)
}
\arguments{
\item{Time}{A POSIXct-format vector or a data.frame containing a similar variable 'Time"}

\item{HHMMSS}{An integer representing time in HHMMSS format}
}
\value{
A numeric index in the Time vector corresponding to the requested time,
or -1 if there is no match.
}
\description{
Returns the index in 'Time' matching HHMMSS
}
\details{
Find the index in a POSIXct time variable that corresponds to a specified time in HHMMSS format.
Returns -1 if the requested time is outside the range in 'Time'
}
\examples{
idx <- getIndex (RAFdata, 201130)
}
\author{
William Cooper
}

