% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRData.R
\name{getRData}
\alias{getRData}
\alias{getRdata}
\title{getRData}
\usage{
getRData(subsetTime = FALSE, newFileName = FALSE)
}
\arguments{
\item{subsetTime}{FALSE by default, which causes all available times to be returned.
If this is TRUE, the user is asked for the time range.}

\item{newFileName}{FALSE by default, in which case the default name and location are
used to store the file (/Data/Work/FD.Rdata). If set TRUE, the user is asked for the
new file name. The data.frame is returned as the value of the function in any case.}
}
\value{
The constructed data.frame, which is also stored as a .RData-format file.
}
\description{
Prepares RAF netCDF data, in data.frame, with user interaction. Returns
it and saves it as an .Rdata file.
}
\details{
For specified variables and time period, constructs a data.frame, saves it 
in Rdata format, and places it as a specified file. Also returns the data.frame
as the value of the function.
}
\author{
William Cooper
}
