% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRoutines.R
\name{getStartEnd}
\alias{GetStartDnd}
\alias{getStartEnd}
\title{getStartEnd}
\usage{
getStartEnd(Time)
}
\arguments{
\item{Time}{A POSIXct format vector or a data.frame containing such a vector named 'Time'.}
}
\value{
c(Start_Time, End_Time), a numeric 2-element vector in HHMMSS format
}
\description{
Find the start and end times of 'Time'
}
\details{
Returns a two-element vector of the start and end times in 'Time'
}
\examples{
getStartEnd (RAFdata)
}
\author{
William Cooper
}

