% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWAC.R
\name{ggplotWAC}
\alias{ggplotWAC}
\title{ggplotWAC}
\usage{
ggplotWAC(.data, col = "blue", xlab = "TIME [UTC]", ylab = "",
  lwd = 1.2, lty = 1, logxy = "", legend.position = "bottomright", ...)
}
\arguments{
\item{.data}{A data.frame containing vectors to plot. The first will be the
abscissa and the remainder ordinate vectors to plot vs the abscissa.}

\item{col}{Color to pass to plot (default: blue). Can be multiple values
to set colors for multiple variables in data.frame mode. There are defaults
('blue', 'darkgreen', 'red', 'skyblue', 'darkorange') but values provided with
this parameter precede that sequence of colors.}

\item{xlab}{Label for the abscissa, to pass to plot (default: "TIME [UTC]")}

\item{ylab}{Label for the ordinate, to pass to plot (default: second
variable name)}

\item{lwd}{Line width to pass to plot (default: 1); can be a vector}

\item{lty}{Line type to pass to plot (default: 1). Can be a vector.
variables provided in data.frame mode.}

\item{logxy}{Set to 'y' for log axes. This is provided (vs log='y') because it
makes possible translation of axis labels to 10^i format.}

\item{legend.position}{This parameter specifies the position for the 
legend. The default is 'bottomright'. The legend
can also be suppressed by setting this parameter to NA. Then a legend can still 
be added after the ggplotWAC call.}

\item{...}{Additional arguments to pass to plot(), but don't include col, xlab, ylab, lwd, type, xaxt or yaxt}
}
\description{
Convenience routine for plots
}
\details{
Sets some plot defaults and calls ggplot and theme_WAC()
}
\examples{
\dontrun{ggplotWAC (data.frame ("Time"=Time, "TASX"=TASX), ylab="TAS")}
\dontrun{plotWAC (Data[, c(Time, PSXCi)], lwd=2)}
\dontrun{plotWAC (Data[, c("Time", "PSXC")])}
}
\author{
William Cooper
}

