% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MurphyKoop.R
\name{GoffGratch}
\alias{GoffGratch}
\alias{goffgratch}
\title{GoffGratch}
\usage{
GoffGratch(Tc, P = 0)
}
\arguments{
\item{Tc}{A numeric vector containing the dew point in deg. C}

\item{P}{An optional numeric vector representing the total pressure in hPa. Set zero 
to suppress the 'enhancement factor' correction for total pressure. Use the pressure to
find the equilibrium water vapor in the presence of that total pressure of air.}
}
\value{
Water vapor pressure [hPa] in equilibrium with a plane water surface at dew point Tc
}
\description{
Returns the water vapor pressure corresponding to a supplied temperature.
}
\details{
Calculates the equilibrium water vapor pressure according to the 
Goff-Gratch equation
}
\examples{
e <- GoffGratch (-12.)
e <- GoffGratch(10., 800.)
EW2 <- GoffGratch (RAFdata$DPXC, RAFdata$PSXC)
}
\author{
William Cooper
}
