% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOAfromRadome.R
\name{GV_YawFromRadome}
\alias{GV_YawFromRadome}
\title{GV_YawFromRadome}
\usage{
GV_YawFromRadome(BDIFR, QCXC)
}
\arguments{
\item{BDIFR}{The pressure difference between starboard and port radome pressures}

\item{QCXC}{The dynamic pressure (use consistent units for BDIFR and QCXC)}
}
\value{
The sideslip angle in degrees (possibly a vector)
}
\description{
Calculate sideslip angle from radome pressures
}
\details{
based on calibration, applies model for sideslip given BDIFR and QCXC
}
\examples{
\dontrun{a <- GV_YawFromRadome (BDIFR, QCXC)}
}
\author{
William Cooper
}

