% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatIndex.R
\name{heatIndex}
\alias{heatIndex}
\alias{heatIndex,}
\alias{HeatIndex,}
\alias{heatindex}
\title{heat index}
\usage{
heatIndex(AT, DPT, RTA = FALSE)
}
\arguments{
\item{AT}{A numeric representing air temperature in deg. C}

\item{DPT}{A numeric representing dewpoint temperature in deg. C.}

\item{RTA}{A logical variable specifying if the apparent temperature
TA should be returned in addition to the heat index. Default is FALSE.}
}
\value{
Either a numeric representing the heat index in deg. C (if
RTA is FALSE) or a named list containing the heat index and the apparent
temperature in deg. C (if RTA is TRUE). Note that
these units are not the conventional ones for the heat index.
}
\description{
Calculates the heat index given the temperature and dewpoint.
}
\details{
This function duplicates the NWS approximating representation
described here: https://www.weather.gov/media/ffc/ta_htindx.PDF .
https://rpubs.com/cooperwilliama/797335 
If the known quantities are temperature and vapor pressure
you can use DPfromE() to find the dewpoint. The
input parameters may be vectors, all of the same size, and in that
case the result will also be a vector. In addition, the function
calculates an apparent temperature described in XXX that
can be used as an alternative to the heat index.
}
\examples{
# expect xx
HI <- Ranadu::heatIndex(20, 10) 
}
\author{
William Cooper
}
