% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCL.R
\name{LCL}
\alias{LCL}
\alias{LiftedCondensationLevel}
\title{LCL}
\usage{
LCL(p, t, r)
}
\arguments{
\item{p}{The pressure [hPa] of the air parcel}

\item{t}{The temperature [degC] of the air parcel}

\item{r}{The water vapor mixing ratio of the air parcel [dimensionless, NOT g/kg]}
}
\value{
A two-element vector giving the pressure [hPa]and temperature [degC] at the LCL
}
\description{
Find the pressure and temperature at the lifted condensation level
}
\details{
Solves iteratively for the pressure at which, for the given 
potential temperature and mixing ratio, the equilibrium vapor pressure at that
pressure level would match that required to give the specified mixing ratio.
Note for comparison the Bolton formula: (2840 / 3.5*log(T+273.15)-log(e)-4.805)+55
}
\examples{
X <- LCL (850, 10, 0.08)  ## expect (825.62, 7.659)
}
\author{
William Cooper
}
