% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWAC.R
\name{lineWAC}
\alias{lineWAC}
\title{lineWAC}
\usage{
lineWAC(x, y, col = "blue", lwd = 2, type = "l", ...)
}
\arguments{
\item{x}{Usually, Time from a data.frame; a vector of abscissa values. Default: Data$Time}

\item{y}{A vector of ordinate values for points to plot.}

\item{col}{Color to pass to plot (default: darkgreen)}

\item{lwd}{Line width to pass to plot (default: 2)}

\item{type}{Line type to pass to plot (default: "l")}

\item{...}{Additional arguments to pass to point()}
}
\description{
Convenience routine for adding lines to plots
}
\details{
Sets some plot defaults and calls points; assumes a plot with axes has already been generated to which to add this line.
}
\examples{
\dontrun{lineWAC (Time, TASX, col='darkgreen')}
\dontrun{lineWAC (Time, PSXC, lty=2)}
}
\author{
William Cooper
}
