% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeNetCDF.R
\name{makeNetCDF}
\alias{makeNetCDF}
\title{makeNetCDF}
\usage{
makeNetCDF(d, newNetCDFname)
}
\arguments{
\item{d}{A data.frame produced by getNetCDF() or otherwise converted to the corresponding
structure from that function.}

\item{newNetCDFname}{A character path name for the file to be created}
}
\value{
none The result is the file on disk
}
\description{
Write a netCDF file using data from a data.frame
}
\details{
This function writes a netCDFfile that includes all variables in a data.frame.
For this to work effectively, the data.frame should be produced by a call to getNetCDF()
because that function also assigns attributes to the data.frame and to the variables it
contains that match those in the original netCDF file. This function extracts those
attributes and adds them to the netCDF file that is created. Note that, when subsetting,
it will be necessary to preserve the attributes; see help for getNetCDF.
}
\examples{
\dontrun{makeNefCDF (Data, "./newFile.nc")}
}
\author{
William Cooper
}

