% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MKerror}
\alias{MKerror}
\title{MKerror}
\usage{
MKerror(DP)
}
\arguments{
\item{DP}{A numeric representing the dew point in deg. C}
}
\value{
The error between the vapor pressure corresponding to the input dew point and the reference vapor pressure (Ereference) set before calling this function.
}
\description{
Internal function used by DPfromE()
}
\details{
Internal function used for inverting the Murphy-Koop formula to find dew-point as a function of water vapor pressure
}
\examples{
\dontrun{err <- MKerror (10.)}
}
\author{
William Cooper
}

