% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncsubset.R
\name{ncsubset}
\alias{ncsubset}
\title{ncsubset}
\usage{
ncsubset(.OldFileName, .NewFileName, .Start = NA, .End = NA,
  .VarList = NULL)
}
\arguments{
\item{.OldFileName}{The path to the existing netCDF file.}

\item{.NewFileName}{The path to the new subset netCDF file. Will be silently removed if present before writing the new file.}

\item{.Start}{The start time (in HHMMSS format) for the new file. Default is the start time for the old file.}

\item{.End}{The end time (in HHMMSS format) for the new file. Default is the end time for the old file.}

\item{.VarList}{A vector of names of variables to include in the new file. Default is the list of variables in the old file.}
}
\value{
NULL; the result is a new file in a specified location.
}
\description{
Make a new netCDF file containing a subset of an existing netCDF file.
}
\details{
A smaller netCDF file is constructed by including only specified variables and a restricted time span
}
\author{
William Cooper
}

