% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotWAC}
\alias{plotWAC}
\title{plotWAC}
\usage{
plotWAC(x, y, col = "blue", xlab = "TIME [UTC]", ylab = "", lwd = 2,
  type = "l", ...)
}
\arguments{
\item{x}{Usually, Time from a data.frame; a vector of abscissa values. Default: Data$Time}

\item{y}{A vector of ordinate values for points to plot.}

\item{col}{Color to pass to plot (default: blue)}

\item{xlab}{Label for the abscissa, to pass to plot (default: "TIME [UTC]")}

\item{ylab}{Label for the ordinate, to pass to plot (default: "")}

\item{lwd}{Line width to pass to plot (default: 2)}

\item{type}{Line type to pass to plot (default: "l")}

\item{...}{Additional arguments to pass to plot(), but don't include col, xlab, ylab, lwd, type, xaxt or yaxt}
}
\description{
Convenience routine for plots
}
\details{
Sets some plot defaults and calls plot and axis
}
\examples{
\dontrun{plotWAC (Time, TASX, ylab="TAS")}
\dontrun{plotWAC (Time, PSXC, lty=2)}
}
\author{
William Cooper
}

