% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotentialTemperatures.R
\docType{data}
\name{potentialTemperatures}
\alias{potentialTemperatures}
\title{potentialTemperatures}
\format{
An object of class \code{list} of length 6.
}
\usage{
potentialTemperatures(DF)  # use $name to get desired function, before (DF)
}
\arguments{
\item{DF}{A data.frame containing PSXC, ATX, EWX, and optionally PLWCD. If the latter
is not present, zero is used for WetEquivalentPotentialTemperature.}
}
\value{
The function value: one of the listed potential temperatures. 
More concise output is obtained from str(lapply(potentialTemperatures, function (f) f(DF), 
digits.d = 5)
}
\description{
List of available potential temperature functions
}
\details{
Contains PotentialTemperature, EquivalentPotentialTemperature,
BoltonEquivalentPotentialTemperature, RossbyEquivalentPotentialTemperature,
VirtualPotentialTemperature, and WetEquivalentPotentialTemperature
in a list that can be used with lapply(potentialTemperatures, function(f) f(DF)) with DF a 
data.frame containing PSXC, ATX, EWX, and optionally PLWCD (for WetEquivalentPotentialTemperature).
To use individual functions: potentialTemperatures$EquivalentPotentialTemperature(DF)
}
\examples{
PTS <- lapply(potentialTemperatures, function(f) 
              f(RAFdata[, c('PSXC', 'ATX', 'EWX')]))
ThetaV <- mean(potentialTemperatures$VirtualPotentialTemperature(
               RAFdata[, c('PSXC', 'ATX', 'EWX')]))
}
\author{
William Cooper
}
\keyword{datasets}
