% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PressureAltitude.R
\name{PressureAltitude}
\alias{PressureAltitude}
\title{PressureAltitude}
\usage{
PressureAltitude(P)
}
\arguments{
\item{P}{A numeric representing pressure in hPa}
}
\value{
The pressure altitude in meters
}
\description{
Pressure altitude in the ISA standard atmosphere
}
\details{
Calculates the altitude in the ISA standard atmosphere corresponding to the specified pressure
}
\examples{
PALT <- PressureAltitude (500.)
}
\author{
William Cooper
}

