% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranadu.R
\docType{package}
\name{Ranadu-package}
\alias{Ranadu}
\alias{Ranadu-package}
\title{Package Ranadu}
\description{
This is a package of programs designed for use with the data archives
that contain measurements from research flights of the NCAR/EOL/RAF
aircraft, normally archived in netCDF format. These routines read those
archives, construct R data.frames containing selected variables, and
provide utility functions for working with the measurements.
}

