% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSpikes.R
\name{removeSpikes}
\alias{removeSpikes}
\alias{RemoveSpikes}
\alias{removespikes}
\title{removeSpikes}
\usage{
removeSpikes(v, sdLimit = 4, runLength = 99, loops = 4)
}
\arguments{
\item{v}{The variable from which to remove spikes.}

\item{sdLimit}{The multiple of the running standard deviation at which to remove
spikes. Default is 4.}

\item{runLength}{The number of measurements over which to calculate the running
mean and standard deviation. The default is 99.}

\item{loops}{The number of times to iterate the calculation.}
}
\value{
The input vector with identified spikes removed and replaced by interpolation.
}
\description{
Identifies spikes and replaces them via interpolation
}
\details{
The function calculates a running mean and standard deviation, then flags
points for which the deviation from the mean exceeds a specified multiplier times
the running standard deviation. It then replaces those points by interpolation.
It repeats this up to a specified number of times, in case the running values are changed by
the removal-and-interpolation process.
}
\examples{
RAFdata$DSPIKED <- removeSpikes(RAFdata$ATTACK)
RAFdata$DSPIKED <- removeSpikes(RAFdata$ATTACK, sdLimit=2, runLength=49, loops=10)
}
\author{
William Cooper
}
