% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsubset.R
\name{Rsubset}
\alias{RSubset}
\alias{Rsubset}
\title{Rsubset}
\usage{
Rsubset(Data, StartT = 0, EndT = 0, Var = NULL, Test = TRUE)
}
\arguments{
\item{Data}{The data.frame from which a subset is extracted. Must contain a
POSIX-format Time variable.}

\item{StartT}{HHMMSS-format starting time, or 0 to start at the first row.}

\item{EndT}{HHMMSS-format ending time, or 0 to end at the last row}

\item{Var}{A character vector of variables to include, or 'ALL' to include
all columns in the data.frame. The default is 'ALL', and Var=NULL produces this
result also.}

\item{Test}{A logical vector used to include rows. The default is TRUE, which will
include all rows. This is provided so that tests like (Data$TASX > 130) can restrict
what is included in the subset data.frame}
}
\value{
A subset data.frame that retains the attributes of the original data.frame.
}
\description{
Construct subset data.frame while preserving attributes
}
\details{
The 'subset()' or '[]' methods remove global attributes from the resulting
data.frame. This function preserves those attributes or, where necessary (e.g., 
Dimensions), changes them appropriately.
}
\examples{
DS <- Rsubset (RAFdata, 201100,201230)

}
\author{
William Cooper
}

