% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTime.R
\name{selectTime}
\alias{selectTime}
\alias{selectTime,}
\alias{selectTimes}
\title{selectTime}
\usage{
selectTime(.d, StartTime, EndTime)
}
\arguments{
\item{.d}{A data.frame or tibble that follows RANADU conventions. It
should contain a POSIXCT-format variable named "Time".}

\item{StartTime}{An initial time in HHMMSS format (e.g., 124500). The
default is 0.}

\item{EndTime}{An ending time in HHMMSS format. Default = 400000.}
}
\value{
A subset tibble or data.frame restricted to the specified 
time range. Limits are inclusive.
}
\description{
Given a RANADU-convention data.frame or tibble, this function
returns a subset in a specified range of times.
}
\details{
The function uses the routine setRange but, instead of returning
a set of indices to use when subsetting the data.frame, it returns a
new subsetted data.frame. This is suitable for use in pipes like
D %>% selectTime(Start, End) %>% dplyr::select(Time, ATX) %>% plotWAC()
}
\author{
William Cooper
}
