% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFileName.R
\name{setFileName}
\alias{setFileName}
\title{setFileName}
\usage{
setFileName(Project = NA, Flight = NA)
}
\arguments{
\item{Project}{The character-name of the project. If provided, the function proceeds to
display the files in that project directory. Otherwise, if "Project" is omitted
or NA (the default) a set of available projects is displayed. Click on one to
select it.}

\item{Flight}{A character name specifying the data file. If omitted, a list
of available files will be displayed from which you can select the file by
clicking on the button. If omitted, '.nc' will be appended to the name.
If "Flight" is specified as, for example, "rf01", the
name will be modified by prefixing the name of the project; e.g., CSETrf01.nc}
}
\value{
A character variable specifying the location of the data.file, 
suitable for use in "getNetCDF()" or "DataFileInfo()".
}
\description{
GUI for selecting from available projects and files
}
\details{
Generates a set of buttons from which to select the project, and 
then another set from which to select the data file. Clicking on the
buttons selects that item.
}
\examples{
\dontrun{setFileName("CSET", "CSETrf05.nc")}
}
\author{
William Cooper
}
