% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRoutines.R
\name{setRange}
\alias{setRange}
\title{setRange}
\usage{
setRange(Time, Start = 0, End = 0)
}
\arguments{
\item{Time}{A POSIXct-format Time variable or a data.frame containing such a variable}

\item{Start}{The desired start time in HHMMSS format (defaults to 0, which gives first index 1)}

\item{End}{The desired end time in HHMMSS format (defaults to the last time in the array)}
}
\value{
A vector of indices corresponding to times in the requested time interval (inclusive).
}
\description{
Set the index range based on requested time limits
}
\details{
For time variable Time, finds indices that match the supplied start and end times and returns the result as a sequence
}
\examples{
r <- setRange (RAFdata, 201100, 201200)
}
\author{
William Cooper
}
