% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothInterp.R
\name{SmoothInterp}
\alias{SmoothInterp}
\alias{smoothInterp}
\title{SmoothInterp}
\usage{
SmoothInterp(.timeSeries, .maxGap = 1000, .Length = 61, .order = 3)
}
\arguments{
\item{.timeSeries}{A numeric vector containing a series to be smoothed.}

\item{.maxGap}{The largest gap across which to interpolate}

\item{.Length}{The length of the segment for Savitzky-Golay filtering.
(Must be odd; will be set odd if supplied as even.) Default: 61}

\item{.order}{The order of the polynomials to be used for filtering.}
}
\value{
The smoothed and filtered series as a vector.
}
\description{
Smooths a time series and interpolates for missing values
}
\details{
Uses Savitzgy-Golay polynomials to smooth the series, after
interpolating to fill missing values and then, if they still exist,
setting remaining missing values to zero.
}
\examples{
DPsmoothed <- SmoothInterp (RAFdata$DPXC)
}
\author{
William Cooper
}
