% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpecificHeats.R
\name{SpecificHeats}
\alias{SpecificHeat}
\alias{SpecificHeats}
\alias{specificheat}
\alias{specificheats}
\title{SpecificHeats}
\usage{
SpecificHeats(EoverP = 0)
}
\arguments{
\item{EoverP}{A numeric giving the ratio of water vapor pressure to total pressure}
}
\value{
an array of dimension (n,3) containing n rows of (Cp, Cv, R), where n is the number of elements in EoverP
}
\description{
Specific heats and gas constant for moist or dry air
}
\details{
Calculates the specific heat at constant pressure, specific heat at constant volume, and gas constant for air having ratio of water vapor pressure to total pressure of the argument EoverP
}
\examples{
CP <- SpecificHeats()    # gives dry-air valuies, single row
CP <- SpecificHeats(6.11/800.)
\dontrun{CP <- (EWX/PSXC) # returns array if EWX and PSXC are vectors}
}
\author{
William Cooper
}

