% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sqs.R
\name{Sqs}
\alias{Sqs}
\title{Sqs}
\usage{
Sqs(W = 1, Nd, T = 5, p = 700)
}
\arguments{
\item{W}{The updraft [m/s]}

\item{Nd}{A data.frame containing the droplet size distribution with column names "n" and "r" 
for number concentration and radius. Units can be either SI or /cm^3 and micrometers.}

\item{T}{The temperature in degrees Celsius.}

\item{p}{The pressure is hPa.}
}
\value{
A two-component vector containing the quasi-steady supersaturation (in fractional units,
not in %) and the time constant for relaxation to the quasi-steady supersaturation.
}
\description{
Calculate the quasi-steady supersaturation and characteristic relaxation time.
}
\details{
See the equations in Politovich and Cooper (1988) and Cooper (1989)
[http://journals.ametsoc.org/doi/pdf/10.1175/1520-0469%281988%29045%3C1651%3AVOTSIC%3E2.0.CO%3B2]
[http://journals.ametsoc.org/doi/pdf/10.1175/1520-0469%281989%29046%3C1301%3AEOVDGH%3E2.0.CO%3B2]
Given the updraft, droplet size distribution, temperature and pressure, calculates the quasi-steady
supersaturation and the associated time constant. Uses a condensation coefficient of 0.04 and
an accommodation coefficient of 1.
}
\examples{
Sqs (W=2., Nd=data.frame(r=2:20, n=c(rep(0,5), rep(50,10), rep(0,4))), T=4.3, p=771)
}
\author{
William Cooper
}
