% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNetCDF.R
\name{standardVariables}
\alias{StandardVariables}
\alias{standardVariables}
\title{standardVariables}
\usage{
standardVariables(list = NULL)
}
\arguments{
\item{list}{An optional list of variable names to add to the standard list}
}
\description{
Standard set of variables, usually used in call to getNetCDF
}
\details{
Sets a standard list of variable names in VarList suitable for use in a call to getNetCDF. Optionally, add "list" to the variables.
The standard variables are ATX, DPXC, EWX, GGALT, LATC, LONC, MACHX, MR, PALT, PSXC, QCXC, TASX, WDC, WSC, WIC.
}
\examples{
standardVariables (c("VEW", "PLWCC"))
}
\author{
William Cooper
}

