% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TellAbout}
\alias{TellAbout}
\alias{tellAbout}
\title{TellAbout}
\usage{
TellAbout(V)
}
\arguments{
\item{V}{A variable that may be scalar, vector, data.frame}
}
\value{
The summary, which is also printed by the function.
}
\description{
Utility to show characteristics of a variable
}
\details{
Prints the variable type, dimensions, and summary.
}
\examples{
\dontrun{TellAbout(V)}
}
\author{
William Cooper
}

