% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_WAC}
\alias{theme_WAC}
\title{theme_WAC}
\usage{
theme_WAC()
}
\value{
A ggplot theme descriptor.
}
\description{
Sets some ggplot2 preferences
}
\details{
Sets some plot defaults according to WAC preferences
}
\examples{
\dontrun{g <- g + theme_WAC ()}
}
\author{
William Cooper
}

