% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTime.R
\name{transferAttributes}
\alias{transferAttributes}
\alias{TransferAttributes,}
\title{transferAttributes}
\usage{
transferAttributes(d, dsub)
}
\arguments{
\item{d}{A data.frame or tibble that follows RANADU conventions.}

\item{dsub}{A second data.frame or tibble, often a subset of the first,
with variables that all appear in the first.}
}
\value{
A modified tibble or data.frame with attributes added to match those
in the original data.frame.
}
\description{
Preserves data.frame attributes.
}
\details{
Given a RANADU-convention reference data.frame or tibble 
and a second (usually subset) data.frame, this function transfers the 
attributes to variables in the second from the same variables in the
first.  This avoids the loss of attributes that often occurs when 
subset data.frames are constructed.
}
\author{
William Cooper
}
