% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TrueAirspeed}
\alias{TrueAirspeed}
\alias{trueairspeed}
\title{TrueAirspeed}
\usage{
TrueAirspeed(MACH, AT, EoverP = 0)
}
\arguments{
\item{MACH}{A numeric representing the Mach number, possible from MachNumber()}

\item{AT}{A numeric representing the air temperature in deg.C}

\item{EoverP}{An optional numeric representing the ratio of water vapor pressure to total pressure}
}
\value{
The true airspeed in m/s
}
\description{
Returns the true airspeed
}
\details{
Calculates the true airspeed from measurements of Mach number, air temperature, and optionally humidity
}
\examples{
TAS <- TrueAirspeed (0.3, 10.)
TAS <- TrueAirspeed (0.5, 0., 6.11/800.)
}
\author{
William Cooper
}

