% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TellAbout.R
\name{ValueOf}
\alias{ValueOf}
\alias{valueOf}
\title{ValueOf}
\usage{
ValueOf(DataFrame, Variable, HHMMSS)
}
\arguments{
\item{DataFrame}{A dataframe containing 'Variable' and also 'Time',
a POSIX-format time variable used for searching for the desired time.}

\item{Variable}{A variable in DataFrame. The variable can be in the 
form DataFrame$TASX or TASX, or "TASX".}

\item{HHMMSS}{A time in hour-minute-second format (e.g., 134513)}
}
\value{
A numeric containing the value of 'Variable' at the specified time.
}
\description{
Returns value of a variable at a specified time
}
\details{
In a dataframe, finds the index corresponding to
the specified time and returns the value of the specified
variable at that index.
}
\examples{
ValueOf (RAFdata, ATX, 201100)
ValueOf (RAFdata, "ATX", 201100)
ValueOf (RAFdata, RAFdata$ATX, 201100)
}
\author{
William Cooper
}
