% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PotentialTemperatures.R
\name{VirtualPotentialTemperature}
\alias{VirtualPotentialTemperature}
\title{VirtualPotentialTemperature}
\usage{
VirtualPotentialTemperature(Tvir, P, E = 0)
}
\arguments{
\item{Tvir}{A numeric representing the virtual temperature in deg. C}

\item{P}{A numeric representing the pressure in hPa}

\item{E}{A numeric representing the water vapor pressure in hPa. This is not used to calculate the virtual temperature but is used for the adiabatic calculation to 1000 hPa total pressure}
}
\value{
A numeric representing the virtual potential temperature in kelvin
}
\description{
Calculates the virtual potential temperature in kelvin
}
\details{
The potential virtual temperature is calculated from the air temperature, pressure, and mixing ratio of water vapor
}
\examples{
THETAV <- VirtualPotentialTemperature (10., 850.)
THETAV <- VirtualPotentialTemperature (20., 900., 12.)
}
\author{
William Cooper
}

