% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viscosity.R
\name{Viscosity}
\alias{Viscosity}
\title{Viscosity}
\usage{
Viscosity(T = 0, p = 1013.25)
}
\arguments{
\item{T}{The air temperature in deg.C. Default: 0}

\item{p}{The air pressure in hPa. Default 1013.25.}
}
\value{
A two-component list containing the dynamic and kinematic viscosity.
}
\description{
Provides the dynamic and kinematic viscosity of dry air in SI units.
}
\details{
The Sutherland formula is used, with appropriate constants for dry air.
The result for dynamic viscosity is assumed independent of pressure. To
calculate kinematic viscosity, the air density is calculated for dry air
from the provided temperature and pressu8re.
}
\examples{
Viscosity (15., 500)
}
\author{
William Cooper
}
