% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TellAbout.R
\docType{data}
\name{WACf}
\alias{WACf}
\title{WACf}
\format{
An object of class \code{list} of length 6.
}
\usage{
WACf(V)  # use [[n]] to select nth function or $name to select function name.
}
\arguments{
\item{V}{A variable that may be scalar, vector, or data.frame column.}
}
\value{
The function value: mean, sd, sdMean, median, kount, or length. kount is
the number of non-missing values in a vector. Values from individual functions
are returned invisibly; i.e., they must be assigned to something as in the examples 
or they will not print. The result of lapply() will still print if not assigned.
More concise output is obtained from STAT <- lapply(); str(STAT). Consider also,
for concise display: dplyr::as_tibble(lapply(WACf, function(f) f(V)))
}
\description{
List of statistical functions
}
\details{
Contains mean, sd, sdMean, median, kount, length with NAs removed, in a list
that can be used with lapply(WACf, function(f) f(V)) with V a vector. To use individual
functions: WACf$sdMean(V). kount is the number of non-missing values in the parameter.
}
\examples{
STAT <- lapply(WACf, function(f) f(RAFdata$TASX))
sdATX <- WACf$sd(RAFdata$ATX)
}
\author{
William Cooper
}
\keyword{datasets}
