% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WetEquivalentPotentialTemperature}
\alias{WetEquivalentPotentialTemperature}
\title{WetEquivalentPotentialTemperature}
\usage{
WetEquivalentPotentialTemperature(P, AT, E, w)
}
\arguments{
\item{P}{A numeric representing pressure in hPa}

\item{AT}{A numeric representing air temperature in deg. C}

\item{E}{A numeric representing water vapor pressure in hPa}

\item{w}{A numeric representing liquid water content in g/m^3}
}
\value{
A numeric representing the wet-equivalent potential temperature in kelvin
}
\description{
The wet-equivalent potential temperature (THETAQ)
}
\details{
Wet-equivalent potential temperature is calculated from the pressure, air temperature, water vapor pressure, and liquid water content; see Paluch 1978.
}
\examples{
THETAQ <- WetEquivalentPotentialTemperature (700., 0., 6.11, 1.0)
}
\author{
William Cooper
}

