% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XformLA.R
\name{XformLA}
\alias{XformLA}
\title{XformLA}
\usage{
XformLA(data, Avector, .inverse = FALSE)
}
\arguments{
\item{data}{A data.frame containing at least ROLL, PITCH, THDG (true
heading) in units of degrees, possibly in many rows representing time
series}

\item{Avector}{A body-frame matrix with three components (x,y,z) in
the aircraft reference frame where x is forward, y is starboard, and
z is downward. For transforming accelerations, body-normal-acceleration
is normally measured in the -z direction in the aircraft frame and has
the acceleration of gravity G subtracted, so G should be added before
transforming and subtracted afterward. The matrix should have 3 columns
representing the components and a number of rows corresponding to the
number of observations.}

\item{.inverse}{Logical, transform from l-frame to b-frame if TRUE}
}
\value{
The vector components transformed to l-frame or ENU coordinates, local
Earth-relative with x east, y north, and z upward. Same structure as Avector.
}
\description{
Transform a vector from b-frame to l-frame.
}
\details{
Apply rotations specified by ROLL, PITCH, THDG in the supplied
data.frame to transform a vector (BX, BY, BZ) in the body frame of the
aircraft (where x is forward, y starboard, z down) to the local or ENU 
frame with x-east, y-north, z-up.
}
\examples{
newDataFrame <- XformLA (data.frame("ROLL"=1:50, "PITCH"=(3+(1:50)/50), "THDG"=91:140),
                                   Avector=matrix(c(21:70, 31:80, 41:90), ncol=3))
}
\author{
William Cooper
}

