% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/XformLB.R
\name{XformLB}
\alias{XformLB}
\title{XformLB}
\usage{
XformLB(data = Data)
}
\arguments{
\item{data}{A data.frame containing at least ROLL, PITCH, THDG (true
heading) in units of degrees, possibly in many rows representing time
series, and a body-frame vector with components BX, BY, BZ.}
}
\value{
The same data.frame with added variables LX, LY, LZ representing
the l-frame coordinates of the vector.
}
\description{
Transform a vector from b-frame to l-frame.
}
\details{
Apply rotations specified by ROLL, PITCH, THDG in the supplied
data.frame to transform a vector (BX, BY, BZ) in the body frame of the
aircraft (where x is forward, y starboard, z down) to the local or ENU
frame with x-east, y-north, z-up.
}
\examples{
newDataFrame <- XformLB (data.frame("ROLL"=1:50, "PITCH"=(3+(1:50)/50), "THDG"=91:140,
                                   "BX"=21:70, "BY"=31:80, "BZ"=41:90))
}
\author{
William Cooper
}

