% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirTemperature.R
\name{AirTemperature}
\alias{AirTemperature}
\title{AirTemperature}
\usage{
AirTemperature(RT, P, Q, E = 0, probe = "HARCO")
}
\arguments{
\item{RT}{A numeric representing the recovery temperature in deg. C}

\item{P}{A numeric representing the ambient pressure in hPa}

\item{Q}{A numeric representing the dynamic pressure in hPa}

\item{E}{An optional numeric representing the water vapor pressure in hPa}

\item{probe}{An identifier for type of temperature probe. Default is 'HARCO'.
Other options are 'ROSE' for heated Rosemount 102,
'HARCOB' for element-B of HARCO, and 'UNHEATED' for unheated Rosemount 102AL.}
}
\value{
The ambient air temperature in deg.C
}
\description{
Calculates the ambient air temperature
}
\details{
Finds the air temperature from measurements of recovery temperature and ambient and dynamic pressure, with optional correction for humidity
}
\examples{
AT <- AirTemperature (10., 700., 50.)
AT <- AirTemperature (RAFdata$RTH1, RAFdata$PSXC, RAFdata$QCXC)
}
\author{
William Cooper
}
