% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcAttack}
\alias{calcAttack}
\title{calcAttack}
\usage{
calcAttack(PITCH, VSPD, TASX)
}
\arguments{
\item{PITCH}{The pitch angle in degrees}

\item{VSPD}{The rate of climb of the aircraft in m/s}

\item{TASX}{The true airspeed of the aircraft in m/s}
}
\description{
Calculate angle-of-attack
}
\details{
Assumes vertical wind = 0, finds AOA from pitch, VSPD and TASX
}
\examples{
\dontrun{a <- calcAttack (PITCH, VSPD, TASX)}
}
\author{
William Cooper
}

