% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcATTACK.R
\name{calcAttack}
\alias{calcAttack}
\title{calcAttack}
\usage{
calcAttack(PITCH, VSPD, TASX)
}
\arguments{
\item{PITCH}{The pitch angle in degrees}

\item{VSPD}{The rate of climb of the aircraft in m/s}

\item{TASX}{The true airspeed of the aircraft in m/s}
}
\description{
Calculate angle-of-attack
}
\details{
Assumes vertical wind = 0, finds AOA from pitch, VSPD and TASX. Used to
provide a reference value for calibrating gust probes. The radome "calibration"
coefficients are usually determined by fitting to reference values provided by
this function, so it should be used where it is expected that the vertical wind
is zero or averages to zero.
}
\examples{
AOAREF <- calcAttack (RAFdata$PITCH, RAFdata$GGVSPD, RAFdata$TASX)
plot (RAFdata$ATTACK, calcAttack (RAFdata$PITCH, RAFdata$GGVSPD, RAFdata$TASX))
}
\author{
William Cooper
}

