% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ComplementaryFilter}
\alias{ComplementaryFilter}
\title{ComplementaryFilter}
\usage{
ComplementaryFilter(FastSignal, SlowSignal, tau = 200)
}
\arguments{
\item{FastSignal}{A numeric vector with a fast-response signal.}

\item{SlowSignal}{A numeric vector that is slower but has less absolute error than the fast-response signal.}

\item{tau}{The time constant for the low-pass filter, in units of the spacing of measurements in the input arrays. (Default: 200)}
}
\value{
The vector of measurements that combines the two input signals.
}
\description{
Calculates a result from combining a fast-response and slow-response signal
}
\details{
Applies a Butterworth filter to calculate the low-pass-filtered result of the difference between the two signals, and adds this filtered result to the signal with fast response.
}
\examples{
\dontrun{ComplementaryFilter(VNS, GGVNS, 150)}
\dontrun{ComplementaryFilter(VEW, GGVEW)}
}
\author{
William Cooper
}

