% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DPfromE}
\alias{DPfromE}
\title{DPfromE}
\usage{
DPfromE(.E)
}
\arguments{
\item{E}{A numeric representing the water vapor pressure in hPa}
}
\value{
A numeric representing the dewpoint in deg. C corresponding to the input vapor pressure.
}
\description{
Finds the dew point given the water vapor pressure
}
\details{
The Murphy-Koop equation for water vapor pressure is inverted to find dew point (in deg. C) from water vapor pressure (in hPa). The inversion uses the R routine nleqslv.
}
\examples{
DP <- DPfromE (0.01)
}
\author{
William Cooper"
}

