% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TellAbout.R
\name{GetAttributes}
\alias{GetAttributes}
\title{GetAttributes}
\usage{
GetAttributes(fname = fname, vname)
}
\arguments{
\item{fname}{The name of the netCDF file.}

\item{vname}{The name of a variable in the netCDF file. This should
be specified as a string variable (e.g., "TASX")}
}
\value{
A list of the names of the attributes
}
\description{
List the netCDF attributes for a specified variable.
}
\details{
Prints the netCDF attributes and their values, and returns
a text vector with the names of the attributes, for a specified
netCDF file and variable. The name of the variable should be
specified as a string.
}
\examples{
\dontrun{attr <- GetAttributes ("/src/raf_data/Projectrf01.nc", "WDC"}
}
\author{
William Cooper
}

