% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIndex}
\alias{getIndex}
\title{getIndex}
\usage{
getIndex(Time, HHMMSS)
}
\arguments{
\item{Time}{A POSIXct-format vector}

\item{HHMMSS}{An integer representing time in HHMMSS format}
}
\value{
A numeric index in the Time vector
}
\description{
Returns the index in 'Time' matching HHMMSS
}
\details{
Find the index in a POSIXct time variable that corresponds to a specified time in HHMMSS format.
}
\examples{
\dontrun{index <- getIndex (Time, 142503)}
}
\author{
William Cooper
}

