% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRAFData.R
\name{getRAFData}
\alias{getRAFData}
\alias{getRAFdata}
\title{getRAFData}
\usage{
getRAFData(.Rdata = "Temp.Rdata.gz", .Directory = "/scr/raf_data", .Project,
  .Flight, .Start = 0, .End = 0, .VarList = NA, .F = 0)
}
\arguments{
\item{.Rdata}{The name to be given to the compressed data file placed on $FTP. Example: "Temp" to get "Temp.Rdata.gz". Default "Temp".}

\item{.Directory}{The root directory for the data, e.g., /scr/raf/ProdData or /scr/raf_data (character string). Default is '/scr/raf_data'.}

\item{.Project}{Project Name (e.g., CONTRAST) -- a directory on /scr/raf_data (character string)}

\item{.Flight}{Flight name (e.g., rf05) (character string)}

\item{.Start}{Start time in HHMMSS format (numeric); default is the first time in the file}

\item{.End}{End time in HHMMSS format (numeric); default is the last time in the file}

\item{.VarList}{List of variables to include in the file (vector of character strings), default is the list of variables provided by 'standardVariables()'.}

\item{.F}{A numeric flight number to include in the data frame as 'RF'. Defaults to 0 and, if 0, is not included.}
}
\value{
The full address where the compressed Rdata file was saved, or NULL in case of failure.
}
\description{
Prepares RAF netCDF data, in data.frame, for FTP download
}
\details{
For specified variables and time period, constructs a data.frame, saves it in Rdata format, gzipped, and places it on $FTP with a supplied name.
}
\author{
William Cooper
}

