% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getStartEnd}
\alias{getStartEnd}
\alias{getstartend}
\title{getStartEnd}
\usage{
getStartEnd(Time)
}
\arguments{
\item{Time}{A POSIXct format vector}
}
\value{
c(Start_Time, End_Time), a numeric 2-element vector in HHMMSS format
}
\description{
Find the start and end times of 'Time'
}
\details{
Returns a two-element vector of the start and end times in 'Time'
}
\examples{
\dontrun{SE <- getStartEnd (Time)}
}
\author{
William Cooper
}

