% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotWAC.R
\name{ggplotWAC}
\alias{ggplotWAC}
\title{ggplotWAC}
\usage{
ggplotWAC(data = Data, xp = Time, yp = ATX, col = "blue",
  xlab = "TIME [UTC]", ylab = "", lwd = 2, type = "l", title = "",
  ...)
}
\arguments{
\item{data}{A dataframe containing the variables to plot}

\item{xp}{Usually, Time from a data.frame; a vector of abscissa values. Default: Data$Time}

\item{yp}{A vector or set of vectors containing ordinate values for points to plot.}

\item{col}{Color or colors to pass to plot (default: blue)}

\item{xlab}{Label for the abscissa, to pass to plot (default: "TIME [UTC]")}

\item{ylab}{Label for the ordinate, to pass to plot (default: "")}

\item{lwd}{Line width to pass to plot (default: 1)}

\item{type}{Line type to pass to plot (default: "l")}

\item{title}{A title for the plot}

\item{...}{Additional arguments to pass to plot(), but don't include col, xlab, ylab, lwd, type, xaxt or yaxt}
}
\value{
A ggplot descriptor.
}
\description{
Convenience routine for plots
}
\details{
Sets some plot defaults and calls ggplot, using
theme_WAC for the plot conventions. This can be followed with
calls to lineWAC and legendWAC to add lines and a legend.
The routine returns a descriptor for the plot that can be
used to repeat the plot with additions. Because the descriptor
that is returned is evaluated in the environment of the calling
program, the calling arguments must be available in that
environment.
}
\examples{
\dontrun{plotWAC (Time, TASX, ylab="TAS")}
\dontrun{plotWAC (Time, PSXC, lty=2)}
}
\author{
William Cooper
}

