% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gravity.R
\name{Gravity}
\alias{Gravity}
\title{Gravity}
\usage{
Gravity(latitude, altitude = 0)
}
\arguments{
\item{latitude}{The latitude (or vector of latitudes) in units of degrees N.}

\item{altitude}{The altitude in meters above the WGS-84 geoid. GPS measurements are usually appropriate to use when supplying this altitude.}
}
\value{
The acceleration of gravidy in units of m/s^2
}
\description{
Provides the acceleration caused by gravity.
}
\details{
The formula used incorporates the dependence on latitude and on altitude above the WGS-84 geoid.
}
\examples{
\dontrun{Gravity (GGLAT[r], GGALT[r])}
}
\author{
William Cooper
}

