% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOAfromRadome.R
\name{GV_AOAfromRadome}
\alias{GV_AOAfromRadome}
\title{GV_AOAfromRadome}
\usage{
GV_AOAfromRadome(ADIFR, QCXC, MACH)
}
\arguments{
\item{ADIFR}{The pressure difference between top and bottom radome ports}

\item{QCXC}{The dynamic pressure (use consistent units for ADIFR and QCXC)}

\item{MACH}{The Mach number}
}
\value{
The angle of attack in degrees (possibly a vector)
}
\description{
Calculate angle of attack from radome pressures
}
\details{
based on calibration, applies model for AOA given ADIFR, QCXC, and Mach Number
}
\examples{
AOA <- GV_AOAfromRadome (RAFdata$ADIFR, RAFdata$QCXC, 
       MachNumber (RAFdata$PSXC, RAFdata$QCXC))
}
\author{
William Cooper
}

