% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histWAC.R
\name{histWAC}
\alias{histWAC}
\title{histWAC}
\usage{
histWAC(
  x,
  Exceedance = FALSE,
  ADD = FALSE,
  breaks = 100,
  freq = FALSE,
  logAxis = "",
  ...
)
}
\arguments{
\item{x}{The usual first argument to hist().}

\item{Exceedance}{Logical (default FALSE). If TRUE, the exceedance is
plotted instead of the conventional cumulative distribution.}

\item{ADD}{Logical (default FALSE) specifying if the distribution should
be added to an existing one.}

\item{breaks}{Number of breaks for histogram. Default = 100.}

\item{freq}{Logical controlling plotting frequency or density. Default FALSE.}

\item{logAxis}{Character vector specifying if the axes should be log. 
Default: FALSE.}

\item{...}{Additional arguments passed to the hist() routine to control 
graphics parameters etc.}
}
\value{
T -- The result is the plot, but the usual value returned by
hist() is also returned invisibly. Assign it to a variable to use it.
}
\description{
Plot a normal histogram but add a cumulative distribution function.
}
\details{
The usual histogram with conventional arguments is used, but
the result then has an added cumulative distribution function with
scale on the right side. The 'density' plot is used and the cumulative
distribution function has limits from 0 to 1.
}
\author{
William Cooper
}
