% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KingProbe.R
\name{KingProbe}
\alias{KingProbe}
\title{KingProbe}
\usage{
KingProbe(Power, TAS, p, T, N)
}
\arguments{
\item{Power}{The power in watts measured by the King probe (Watts)}

\item{TAS}{The true airspeed (m/s)}

\item{p}{Pressure (hPa)}

\item{T}{Temperature (deg. C)}

\item{N}{Droplet concentration (e.g., from a CDP)}
}
\value{
Liquid water content (g/m^3)
}
\description{
Calculates liquid water content from King Probe
}
\details{
Calculates an estimate of the dry-probe power and
subtracts this from the total power, then converts the resulting
residual power to a measure of the liquid water content. If applied
to successive measurements, the routine maintains an estimate of
the dry-air relationship between Nusselt number and Reynolds number
and uses that to adjust the zero reading when out of cloud (as
indicated by a measured droplet concentration falling below some
threshold).
}
\examples{
\dontrun{lwc <- KingProbe (25, 180, 700, 10, 0) }
\dontrun{lwc <- KingProbe (RAFdata$PLWC, RAFdata$TASX, RAFdata$PSXC, 
RAFdata$ATX, RAFdata$CONCD_LWII)}
}
\author{
William Cooper
}

