% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MurphyKoop.R
\name{MurphyKoop}
\alias{MurphyKoop}
\alias{murphykoop}
\title{MurphyKoop}
\usage{
MurphyKoop(DP, P = 0)
}
\arguments{
\item{DP}{A numeric representing the dew point in deg. C}

\item{P}{An optional numeric representing total pressure in hPa, set zero to suppress 'enhancement factor' correction for total pressure}
}
\value{
Water vapor pressure [hPa] in equilibrium with a plane water surface at dew point DP
}
\description{
Returns the water vapor pressure
}
\details{
Calculates the equilibrium water vapor pressure according to the Murphy-Koop equation
}
\examples{
e <- MurphyKoop (-12.)
e <- MurphyKoop (10., 800.)
}
\author{
William Cooper
}

