% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MurphyKoop.R
\name{MurphyKoopIce}
\alias{MurphyKoopIce}
\alias{murphykoopice}
\title{MurphyKoopIce}
\usage{
MurphyKoopIce(FP, P = 0)
}
\arguments{
\item{FP}{A numeric representing the frost point in deg. C}

\item{P}{An optional numeric representing total pressure in hPa, set zero to 
suppress 'enhancement factor' correction for total pressure}
}
\value{
Water vapor pressure in equilibrium with a plane ice surface at frost point FP
}
\description{
Returns the water vapor pressure
}
\details{
Calculates the water vapor pressure in equilibrium over a plane ice surface according to the Murphy-Koop equation
}
\examples{
e <- MurphyKoopIce (-50.)
e <- MurphyKoopIce (-60., 200.)
}
\author{
William Cooper
}

