% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCorFunction.R
\name{PCorFunction}
\alias{PCorFunction}
\alias{PCORfunction}
\title{Static Defect Correction}
\usage{
PCorFunction(p, q = 100, akrd = 2.5, AC = "GV", Sensors = "STD")
}
\arguments{
\item{p}{Measured (uncorrected) ambient pressure (e.g., PSF), hPa (no default)}

\item{q}{Measured (uncorrected) dynamic pressure (e.g., QCF), hPa (or any units 
that match those used for p) -- default 100 hPa.}

\item{akrd}{The angle of attack (degrees; 2.5 deg. default)}

\item{AC}{Aircraft identifier, either "GV" of "C130" (default "GV")}

\item{Sensors}{Normally "STD" (the default) but "R" for right pressure pair, C130}
}
\value{
The ratio of the pressure correction to the measured pressure. The 
correction to dynamic pressure is the negative of the returned value.
}
\description{
Corrections to dynamic pressure and ambient pressure
}
\details{
This correction was determined from fits to the measurements from
the Laser Air Motion Sensor, as described by Cooper et al., 2014. These do not
apply to older projects; in particular, TREX and before (GV).
}
\examples{
# using PSXC and QCXC because PSF and QCF are not in RAFdata
PCOR <- RAFdata$PSXC * PCorFunction (RAFdata$PSXC, RAFdata$QCXC, RAFdata$ATTACK)
\dontrun{dPoverP <- PCorFunction (p, q, akrd, "GV")}
}
\author{
William Cooper
}
