% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PCorFunction}
\alias{PCorFunction}
\title{Static Defect Correction}
\usage{
PCorFunction(p, q = 100, akrd = 2.5, AC = "GV")
}
\arguments{
\item{p}{Measured (uncorrected) ambient pressure (e.g., PSF), hPa (no default)}

\item{q}{Measured (uncorrected) dynamic pressure (e.g., QCF), hPa (or any units that match those used for p) -- default 100 hPa. This is unused for the C-130 so it does not need to be supplied.}

\item{akrd}{Ther angle of attack (degrees; no default)}

\item{AC}{Aircraft identifier, either "GV" of "C130" (default "GV")}
}
\value{
The ratio of the pressure correction to the measured pressure. The correction to dynamic pressure is the negative of the returned value.
}
\description{
Corrections to dynamic pressure and ambient pressure
}
\details{
This correction was determined from fits to the measurements from
the Laser Air Motion Sensor, as described by Cooper et al., 2014.
}
\examples{
\dontrun{dPoverP <- PCorFunction (p, q, akrd, "GV")}
}
\author{
William Cooper
}

