% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotWAC.R
\name{plotWAC}
\alias{plotWAC}
\title{plotWAC}
\usage{
plotWAC(x, y = NA, col = "blue", xlab = "TIME [UTC]", ylab = "",
  lwd = 2, type = "l", legend.position = NA, ...)
}
\arguments{
\item{x}{Usually, Time from a data.frame; a vector of abscissa values. Optionally,
a data.frame containing multiple variables to plot and superceding parameter y.}

\item{y}{A vector of ordinate values for points to plot. If the first parameter
is a data.frame, this argument is ignored.}

\item{col}{Color to pass to plot (default: blue)}

\item{xlab}{Label for the abscissa, to pass to plot (default: "TIME [UTC]")}

\item{ylab}{Label for the ordinate, to pass to plot (default: "")}

\item{lwd}{Line width to pass to plot (default: 2)}

\item{type}{Line type to pass to plot (default: "l")}

\item{legend.position}{For multiple-line calls with a data.frame, this parameter
specifies the position for the legend. The default is to produce no legend, but
this can still be added after the plotWAC call.}

\item{...}{Additional arguments to pass to plot(), but don't include col, xlab, ylab, lwd, type, xaxt or yaxt}
}
\description{
Convenience routine for plots
}
\details{
Sets some plot defaults and calls plot and axis
}
\examples{
\dontrun{plotWAC (Time, TASX, ylab="TAS")}
\dontrun{plotWAC (Time, PSXC, lty=2)}
}
\author{
William Cooper
}

