% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PressureAltitude.R
\name{PressureAltitude}
\alias{PressureAltitude}
\alias{pressureAltitude}
\title{PressureAltitude}
\usage{
PressureAltitude(P)
}
\arguments{
\item{P}{A numeric vector representing pressure in hPa}
}
\value{
A numeric vector giving the pressure altitude in meters
}
\description{
Pressure altitude in the ISA standard atmosphere
}
\details{
Calculates the altitude in the ISA standard atmosphere corresponding to the specified pressure
}
\examples{
PALT <- PressureAltitude (500.)
PALT2 <- PressureAltitude (RAFdata$PSXC)
}
\author{
William Cooper
}

