% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RecoveryFactor}
\alias{RecoveryFactor}
\alias{recoveryfactor}
\title{RecoveryFactor}
\usage{
RecoveryFactor(MACH)
}
\arguments{
\item{MACH}{A numeric giving the Mach number, perhaps calculated by MachNumber()}
}
\value{
The recovery factor applicable to a heated sensor
}
\description{
Mach-number-dependent recovery factor for heat sensor
}
\details{
For specified Mach number, returns the recovery factor according to the formula in ProcessingAlgorithms.
}
\examples{
rf <- RecoveryFactor (0.75)
}
\author{
William Cooper
}

