% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RecoveryFactor.R
\name{RecoveryFactor}
\alias{RecoveryFactor}
\alias{recoveryfactor}
\title{RecoveryFactor}
\usage{
RecoveryFactor(.M, probe = "HARCO")
}
\arguments{
\item{probe}{A string giving a probe name. Default is 'HARCO'; other options are
'HARCOB' for B-element of HARCO, 'Rose" (heated 102 Rosemount) and 'UNHEATED' (Rosemount 102AL unheated)}

\item{MACH}{A numeric giving the Mach number, perhaps calculated by MachNumber()}
}
\value{
The recovery factor applicable to a temperature sensor
}
\description{
Mach-number-dependent recovery factor for a temperature sensor
}
\details{
For specified Mach number, returns the recovery factor according to the formula in ProcessingAlgorithms.
}
\examples{
rf <- RecoveryFactor (0.75)
}
\author{
William Cooper
}

