% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFileName.R
\name{setFileName}
\alias{setFileName}
\alias{setVariables}
\title{setVariableList}
\usage{
setFileName(Project = NA, Flight = NA)
}
\arguments{
\item{fname}{The name of a netCDF file containing variables.}

\item{VarList}{A vector of character names to be selected before user interaction.
This might be the result of a call to Ranadu::standardVariables(), for example.}

\item{single}{A logical flag indicating if only one selection should be returned and
the function should return immediately after the first selection is made.}
}
\value{
A vector of character names of the selected variables.
}
\description{
GUI for selecting from available netCDF variables (under development)
}
\details{
A display of buttons for all the available variables in a netCDF
file is generated, with variables already in 'VarList' highlighted.
Clicking on buttons highlights them also, and when the File->Set Selections
menu item is selected the highlighted list is written to 'VarList'. This
routine does some dangerous things with scope to make the tcltk structure
work, including assuming that 'VarList' exists in the global environment
and, at the end, writing to that variable list. It will over-write an 
existing 'VarList'.
}
\examples{
\dontrun{setVariableList("/scr/raf_data/WINTER/WINTERrf11.nc", c("ATX", "PSXC"))}
}
\author{
William Cooper
}
