% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TellAbout.R
\name{TellAbout}
\alias{TellAbout}
\alias{tellAbout}
\title{TellAbout}
\usage{
TellAbout(V)
}
\arguments{
\item{V}{A variable that may be scalar, vector, data.frame}
}
\value{
The summary, which is also printed by the function.
}
\description{
Utility to show characteristics of a variable
}
\details{
Prints the variable type, dimensions, and summary.
}
\examples{
\dontrun{TellAbout("TASX")}
}
\author{
William Cooper
}

