% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_WAC.R
\name{theme_WAC}
\alias{theme_WAC}
\title{theme_WAC}
\usage{
theme_WAC(version = 0)
}
\arguments{
\item{version}{The style number (default 0); 1 used for faceted 
plots and other adaptations in KalmanFilterTechNote, e.g.}
}
\value{
A ggplot theme descriptor.
}
\description{
Sets some ggplot2 preferences
}
\details{
Sets some plot defaults according to WAC preferences
}
\examples{
\dontrun{g <- g + theme_WAC ()}
}
\author{
William Cooper
}
