% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrueAirspeed.R
\name{TrueAirspeed}
\alias{TrueAirspeed}
\alias{trueAirspeed}
\title{TrueAirspeed}
\usage{
TrueAirspeed(MACH, AT, EoverP = 0)
}
\arguments{
\item{MACH}{A numeric vector representing the Mach number, possibly from MachNumber()}

\item{AT}{A numeric vector representing the air temperature in deg.C}

\item{EoverP}{An optional numeric vector representing the ratio of water vapor pressure to total pressure}
}
\value{
A numeric vector giving the true airspeed in m/s
}
\description{
Returns the true airspeed
}
\details{
Calculates the true airspeed from measurements of Mach number, air temperature, and optionally humidity
}
\examples{
TAS <- TrueAirspeed (MachNumber (RAFdata$PSXC, RAFdata$QCXC), RAFdata$ATX, RAFdata$EWX/RAFdata$PSXC)
}
\author{
William Cooper
}

