% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ValueOf}
\alias{ValueOf}
\title{ValueOf}
\usage{
ValueOf(Variable, HHMMSS, DataFrame = Data)
}
\arguments{
\item{Variable}{A variable in a dataframe named Data that also includes Time. The
variable can be in the form Data$TASX or, if Data is attached,
TASX.}

\item{HHMMSS}{A time in hour-minute-second format (e.g., 134513)}

\item{DataFrame}{The dataframe containing Variable}
}
\value{
The value of the supplied variable at the specified time.
}
\description{
Returns value of a variable at a specified time
}
\details{
In a dataframe, finds the index corresponding to
the specified time and returns the value of the specified
variable at that index.
}
\examples{
\dontrun{x <- ValueOf (ATX, 140233)}
}
\author{
William Cooper
}

