% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TellAbout.R
\name{ValueOfAll}
\alias{ValueOfAll}
\title{ValueOfAll}
\usage{
ValueOfAll(DataFrame = Data, HHMMSS)
}
\arguments{
\item{DataFrame}{The specified dataframe, which must contain Time as
a POSIXgt variable.}

\item{HHMMSS}{The time, in hour-minute-second format (e.g., 140321)
at which to report the variables in the dataframe.}
}
\value{
A vector containing all the variables at the specified
time. These values are also printed.
}
\description{
See/Get values of dataframe variables at specified time.
}
\details{
Returns and prints values of all variables in a specified
dataframe that includes Time, at the time specified.
}
\examples{
\dontrun{vlist <- ValueOfAll (Data, 143329)}
}
\author{
William Cooper
}

