% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotentialTemperatures.R
\name{WetEquivalentPotentialTemperature}
\alias{WetEquivalentPotentialTemperature}
\title{WetEquivalentPotentialTemperature}
\usage{
WetEquivalentPotentialTemperature(P, AT, E = 0, w = 0)
}
\arguments{
\item{P}{A numeric representing pressure in hPa}

\item{AT}{A numeric representing air temperature in deg. C}

\item{E}{A numeric representing water vapor pressure in hPa. Default is 0, and if
E <= 0 then E is set to the equilibrium value for the calculation.}

\item{w}{A numeric representing liquid water content in g/m^3. Default is 0.}
}
\value{
A numeric representing the wet-equivalent potential temperature in kelvin
}
\description{
The wet-equivalent potential temperature (THETAQ)
}
\details{
Wet-equivalent potential temperature is calculated from the pressure, air temperature, water vapor pressure, and liquid water content; see Paluch 1978.
}
\examples{
THETAQ <- WetEquivalentPotentialTemperature (700., 0., 6.11, 1.0)
}
\author{
William Cooper
}

